feat: Modulize sending code email #273
