+++ server/controllers/users.ts
@@ -10,7 +10,6 @@ import {
   generateRandomNumber,
 } from "../utils/randomString";
 import { client } from "../_redisConfig";
-import { sendEmail } from "../utils/email";
 import { cipher, decipher } from "../utils/crypto";
 
 import { logger } from "@logger";
